# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/expenses": {
      "get": {
        "operationId": "ExpensesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "지출 목록 [대표]",
        "tags": [
          "expenses"
        ]
      },
      "post": {
        "operationId": "ExpensesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "지출 요청 생성 [대표]",
        "tags": [
          "expenses"
        ]
      }
    },
    "/api/expenses/{id}": {
      "get": {
        "operationId": "ExpensesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "지출 상세 [대표]",
        "tags": [
          "expenses"
        ]
      },
      "patch": {
        "operationId": "ExpensesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExpenseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "지출 수정(requested만 — 오기입 정정) [대표]",
        "tags": [
          "expenses"
        ]
      },
      "delete": {
        "operationId": "ExpensesController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "지출 철회(soft delete, requested만) [대표]",
        "tags": [
          "expenses"
        ]
      }
    },
    "/api/expenses/{id}/approve": {
      "post": {
        "operationId": "ExpensesController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "지출 승인 + 통합 원장 출금 기록 [대표]",
        "tags": [
          "expenses"
        ]
      }
    },
    "/api/expenses/{id}/reject": {
      "post": {
        "operationId": "ExpensesController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectExpenseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ],
        "summary": "지출 반려(사유 필수) [대표]",
        "tags": [
          "expenses"
        ]
      }
    },
...
    "/api/parents": {
      "get": {
        "operationId": "ParentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Parent[] — name·phone·kakaoAvailable"
          }
        },
        "summary": "보호자 목록(Parent[]) — 연락처 포함이라 관리자 전용(강사 403, P0-5) [매니저 이상]",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "post": {
        "operationId": "ParentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "{ parent, relation }"
          }
        },
        "summary": "보호자 등록 + 학생 연결 — studentId FK 검증, 대표 불변",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/parents/relations": {
      "get": {
        "operationId": "ParentsController_findAllRelations",
        "parameters": [],
        "responses": {
          "200": {
            "description": "ParentStudent[] — parentId·studentId·relation·isPayer·isPrimary"
          }
        },
        "summary": "학생↔보호자 관계(ParentStudent[]) — M:N(대표/납부자) [매니저 이상]",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/parents/{id}": {
      "get": {
        "operationId": "ParentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "보호자 단건과 학생 연결 정보 조회 [매니저 이상]",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "patch": {
        "operationId": "ParentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "보호자 연락처·관계 정보를 수정하고 이력 기록 [매니저 이상]",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "delete": {
        "operationId": "ParentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "참조 무결성 확인 후 보호자 soft delete [매니저 이상]",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/parents/link": {
      "post": {
        "operationId": "ParentsController_link",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkParentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 ParentStudent"
          }
        },
        "summary": "기존 보호자를 학생에 연결(형제 M:N) — FK·유니크·대표 불변",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/parents/relations/{id}": {
      "patch": {
        "operationId": "ParentsController_updateRelation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRelationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정된 ParentStudent"
          }
        },
        "summary": "관계 수정(대표 이전·납부자 변경) — 대표 지정 시 기존 대표 강등",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "delete": {
        "operationId": "ParentsController_removeRelation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "학생과 보호자의 연결 관계 soft delete [매니저 이상]",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/parents/relations/{id}/guardian": {
      "delete": {
        "operationId": "ParentsController_removeGuardian",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "학생 연결과 미참조 보호자를 함께 soft delete [매니저 이상]",
        "tags": [
          "parents"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TACO ERP API",
    "description": "TnAcademy 백오피스 운영 API. Postgres 영속 저장소와 역할 기반 접근 제어를 사용하는 실제 서비스 계약.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "현재 호스트"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "access_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "CreateExpenseDto": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "supplies",
              "equipment",
              "books",
              "rent",
              "utility",
              "marketing",
              "meal",
              "etc"
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 100
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "spentAt": {
            "type": "string",
            "maxLength": 30
          },
          "vendor": {
            "type": "string",
            "maxLength": 100
          },
          "memo": {
            "type": "string",
            "maxLength": 500
          },
          "receiptUrl": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "category",
          "title",
          "amount",
          "spentAt"
        ]
      },
      "UpdateExpenseDto": {
        "type": "object",
        "properties": {
          "category": {
            "enum": [
              "supplies",
              "equipment",
              "books",
              "rent",
              "utility",
              "marketing",
              "meal",
              "etc"
            ],
            "type": "string"
          },
          "title": {
            "type": "string",
            "maxLength": 100
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "spentAt": {
            "type": "string",
            "maxLength": 30
          },
          "vendor": {
            "type": "string",
            "maxLength": 100
          },
          "memo": {
            "type": "string",
            "maxLength": 500
          },
          "receiptUrl": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "RejectExpenseDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 200,
            "example": "증빙 누락 — 영수증 첨부 후 재신청",
            "description": "반려 사유(필수)"
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "CreateParentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "pattern": "\\S"
          },
          "phone": {
            "type": "string",
            "maxLength": 30
          },
          "webId": {
            "type": "string",
            "maxLength": 50
          },
          "relation": {
            "type": "string",
            "maxLength": 30
          },
          "isPayer": {
            "type": "boolean"
          },
          "isPrimary": {
            "type": "boolean"
          },
          "studentId": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "name",
          "studentId"
        ]
      },
      "LinkParentDto": {
        "type": "object",
        "properties": {
          "parentId": {
            "type": "number",
            "minimum": 1
          },
          "studentId": {
            "type": "number",
            "minimum": 1
          },
          "relation": {
            "type": "string",
            "maxLength": 30
          },
          "isPayer": {
            "type": "boolean"
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "required": [
          "parentId",
          "studentId"
        ]
      },
      "UpdateRelationDto": {
        "type": "object",
        "properties": {
          "relation": {
            "type": "string",
            "maxLength": 30
          },
          "isPayer": {
            "type": "boolean"
          },
          "isPrimary": {
            "type": "boolean"
          }
        }
      },
      "UpdateParentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "pattern": "\\S"
          },
          "phone": {
            "type": "string",
            "maxLength": 20
          },
          "kakaoAvailable": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "access_token": []
    }
  ]
}